EoDC: Improve kinematic fits for ic1459
errors2_muse.py: altered the mask for ic1459 a little to aid the fitting of Halpha and [NII] lines. This is an atempt to improve the gas kinematic maps. The gas intensity maps looked resonable, but the kinematics were definately wrong... <- odd: a bad fit should be bad in all dimensions of the phase-space.

All other routines: General use

 On branch master
 Your branch is up-to-date with 'joshwarren/master'.

 Changes to be committed:
	modified:   create_params_file_muse.py
	modified:   errors2_muse.py
	modified:   full_analysis2_muse.py
	modified:   full_analysis_muse.py
	modified:   plot_results_muse.py