EoDC: Started stellar pop work and changed directory archetecture.
REMOVED: errors3_muse.py: This has now been completely absorbed into errors2_muse.py

New routine: pop_muse.py: This routine is to fit the LICK indices for the TMJ models to the observations as corrected for kinematics using the opt keyword as 'pop' in errors2_muse.py.

Other routines: Slightly altered directory archetecture: the entire of one run of analysis is placed within a single directory within the relevent galaxy's folder. This means it is easier to keep track of multiple runs e.g. differing levels of binning. The plots should be placed within the plot directory within this folder and the pickled file, in the pickled directory.

kinematics_muse.py: I have uncommented the lambda_R section as this is the next thing to do. This needed redoing: particularly focusing on definitions and find lambda_R_e (not lambda_R_max).

 On branch master
 Your branch is up-to-date with 'joshwarren/master'.

 Changes to be committed:
	modified:   BPT.py
	modified:   binning_spaxels_muse.py
	modified:   create_params_file_muse.py
	modified:   errors2_muse.py
	deleted:    errors3_muse.py
	modified:   full_analysis2_muse.py
	modified:   full_analysis_muse.py
	modified:   kinematics_muse.py
	modified:   pickler_muse.py
	modified:   plot_results_muse.py
	new file:   pop_muse.py