EoDC: Working code!
KDC_pop.py: Working code to compare the inner 1 arcsec stellar populations with the rest of the galaxy.

errors2_muse.py: successfully split up the code into two parts to be able to pass a spectrum into the ppxf wrapper.

pop_muse.py: altered to be able to run with a given ppxf object output.

 On branch master
 Your branch is up-to-date with 'joshwarren/master'.

 Changes to be committed:
	modified:   KDC_pop.py
	modified:   errors2_muse.py
	modified:   pop_muse.py